# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tres Service API Documentation",
        "description": "API documentation for Tres Service application",
        "contact": {
            "email": "sibomanaedouard974@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/folder/new": {
            "post": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Create a new folder",
                "description": "Creates a new folder for the authenticated user",
                "operationId": "createFolder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "New Project Folder"
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ],
                                        "example": "private"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepass123",
                                        "nullable": true
                                    },
                                    "allow_download": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "watermark_images": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Folder created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Folder created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Folder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/folder/upload": {
            "post": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Upload files into a folder",
                "description": "Uploads one or more files to a specified folder",
                "operationId": "uploadFilesInFolder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "folder_id",
                                    "files"
                                ],
                                "properties": {
                                    "folder_id": {
                                        "description": "ID of the target folder",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "files[]": {
                                        "description": "Array of files to upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Files uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Files uploaded successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/File"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Folder not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/folder/{id}": {
            "get": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Get folder details & list files/subfolders",
                "description": "Retrieves details of a specific folder and its contents (files and subfolders)",
                "operationId": "getFolderDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the folder",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Folder details retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "folder": {
                                                    "$ref": "#/components/schemas/Folder"
                                                },
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/File"
                                                    }
                                                },
                                                "subfolders": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Folder"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Folder not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Delete folder (move to trash)",
                "description": "Moves a folder and its contents to the trash using soft delete",
                "operationId": "deleteFolder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the folder",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder moved to trash successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Folder moved to trash successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Folder not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Folder Management"
                ],
                "summary": "Update folder info",
                "description": "Updates properties of a specific folder",
                "operationId": "updateFolder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the folder",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Renamed Folder",
                                        "nullable": true
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private"
                                        ],
                                        "example": "public",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpass123",
                                        "nullable": true
                                    },
                                    "allow_download": {
                                        "type": "boolean",
                                        "example": false,
                                        "nullable": true
                                    },
                                    "watermark_images": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Folder updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Folder updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Folder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Folder not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Folder": {
                "title": "Folder",
                "description": "Folder model",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "privacy",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Owner User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "parent_id": {
                        "description": "Parent folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": null,
                        "nullable": true
                    },
                    "name": {
                        "description": "Folder name",
                        "type": "string",
                        "example": "My Documents"
                    },
                    "privacy": {
                        "description": "Folder privacy",
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "password_protected": {
                        "description": "Is folder password protected?",
                        "type": "boolean",
                        "example": false
                    },
                    "allow_download": {
                        "description": "Allow download from this folder",
                        "type": "boolean",
                        "example": true
                    },
                    "watermark_images": {
                        "description": "Watermark images in this folder",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp (for soft delete)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}